requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cycler==0.10.0
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.42.0
django-ckeditor==5.9.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-password-reset==2.0
django-simpleui==4.0.2
django-taggit==1.2.0
idna==2.9
importlib-metadata==1.6.0
Jinja2==2.11.2
kiwisolver==1.2.0
legacy==0.1.6
MarkupSafe==1.1.1
matplotlib==3.2.1
mpmath==1.1.0
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.4
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
pywordcloud==0.2.2
requests==2.23.0
requests-oauthlib==1.3.0
simpleui==4.0.2
six==1.15.0
sqlparse==0.3.1
sympy==1.5.1
urllib3==1.25.9
wordcloud==1.7.0
zipp==3.1.0
